feat(parser): add parser module and refactor rule construction logic
- Introduce `parser::mod.rs` as the new entrypoint for parsing Semgrep-compatible YAML rules.
- Move mode-specific rule builder functions and validation logic to `parser::builders.rs`.
- Implement detailed validation of principal fields for different rule modes (search, extract, join, taint).
- Add stricter validation for required fields, e.g., non-empty `languages` list in search and extract rules.
- Improve diagnostic error reporting with more precise schema validation.
- Add `Display` implementation for `RuleMode`.
- Include tests validating rejection of invalid rules, cross-mode field misuse, and empty language lists.

This change enhances code organization, improves validation strictness, and prepares for easier maintenance and extension of the YAML rule parsing system.
